pagebench: factor out the concept of thread local stats
