+++ pageserver/pagebench/src/util.rs
@@ -1,3 +1,6 @@
 pub(crate) mod connstring;
 pub(crate) mod discover_timelines;
+pub(crate) mod request_stats;
 pub(crate) mod tenant_timeline_id;
+#[macro_use]
+pub(crate) mod tokio_thread_local_stats;
